logs in and creates user
