+++ App/navigation/AppNavigator.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 import LoginScreen from '../screens/Login/LoginScreen';
 import SignUpScreen from '../screens/Login/SignUpScreen';
+import LandingScreen from '../screens/Login/LandingScreen';
 import MainTabNavigator from './MainTabNavigator';
 import Loading from '../screens/Login/Loading';
 
@@ -15,12 +16,13 @@ export default createAppContainer(
   createSwitchNavigator(
     {
       Loading: Loading,
+      LandingScreen: LandingScreen,
       Login: LoginScreen,
-      Signup: SignUpScreen,
+      SignUp: SignUpScreen,
       Main: MainTabNavigator
     },
     {
-      initialRouteName: 'Loading'
+      initialRouteName: 'LandingScreen'
     }
   )
 );
