+++ App/navigation/MainTabNavigator.js
@@ -9,10 +9,10 @@ import HomeScreen from '../screens/HomeScreens/HomeScreen';
 import Standings from '../screens/StandingsScreen/Standings';
 import ScoresScreen from '../screens/ScoresScreen/ScoresScreen';
 import SplatterScreen from '../screens/SplatterScreen/SplatterScreen';
-import FakeSignUp from '../screens/Login/FakeSignup';
-import FakeLogin from '../screens/Login/FakeLogin';
-import FakeChooseIcon from '../screens/Login/FakeChooseIcon';
-import FakeLandingScreen from '../screens/Login/FakeLandingScreen';
+// import FakeSignUp from '../screens/Login/FakeSignup';
+// import FakeLogin from '../screens/Login/FakeLogin';
+import ChooseIcon from '../screens/Login/ChooseIcon';
+// import FakeLandingScreen from '../screens/Login/FakeLandingScreen';
 import LeagueSelectionScreen from '../screens/LeagueAdminScreens/LeagueSelectionScreen';
 
 import LoginScreen from '../screens/Login/LoginScreen';
@@ -82,10 +82,10 @@ SplatterStack.navigationOptions = {
   )
 };
 const TheRipeStack = createStackNavigator({
-  FakeLandingScreen: FakeLandingScreen,
-  FakeSignUp: FakeSignUp,
-  FakeLogin: FakeLogin,
-  FakeChooseIcon: FakeChooseIcon,
+  // FakeLandingScreen: FakeLandingScreen,
+  // FakeSignUp: FakeSignUp,
+  // FakeLogin: FakeLogin,
+  ChooseIcon: ChooseIcon,
   LeagueSelectionScreen: LeagueSelectionScreen
 });
