+++ App/screens/HomeScreens/HomeScreen.js
@@ -17,27 +17,27 @@ import CurrentVideos from './CurrentVideos';
 import firebase from '../../.././firebase';
 import db from '../../.././db';
 
-function testDB() {
-  firebase.auth().onAuthStateChanged(user => {
-    if (user != null) {
-      Alert.alert('this is state in home screen', user.uid);
-      db.collection('users')
-        .add({
-          first: user.uid,
-          last: user.email,
-          born: 1815
-        })
-        .then(function(docRef) {
-          console.log('Document written with ID: ', docRef.id);
-        })
-        .catch(function(error) {
-          console.error('Error adding document: ', error);
-        });
-    }
-  });
-}
+// function testDB() {
+//   firebase.auth().onAuthStateChanged(user => {
+//     if (user != null) {
+//       Alert.alert('this is state in home screen', user.uid);
+//       db.collection('users')
+//         .add({
+//           first: user.uid,
+//           last: user.email,
+//           born: 1815
+//         })
+//         .then(function(docRef) {
+//           console.log('Document written with ID: ', docRef.id);
+//         })
+//         .catch(function(error) {
+//           console.error('Error adding document: ', error);
+//         });
+//     }
+//   });
+// }
 
-testDB();
+// testDB();
 
 export default class HomeScreen extends React.Component {
   state = { currentUser: null };
