+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "firebase": "^6.0.2",
+    "moment": "^2.24.0",
     "native-base": "^2.12.1",
     "react": "16.8.3",
     "react-native": "^0.59.2",
@@ -29,7 +30,10 @@
     "react-native-svg-charts": "^5.2.0",
     "react-native-vector-icons": "^6.4.2",
     "react-native-youtube": "^1.1.0",
-    "react-navigation": "^3.5.1"
+    "react-navigation": "^3.5.1",
+    "react-redux": "^7.1.0",
+    "redux": "^4.0.1",
+    "redux-logger": "^3.0.6"
   },
   "devDependencies": {
     "@babel/core": "^7.4.0",
